Move CAP commands to irc_cap.c; use enum for flags
